Refactor folder structure, avoid openapi ref syntax.
